// client/src/types/api.schema.ts
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  CountryCode: { input: any; output: any; }
  DateTime: { input: any; output: any; }
  ErrorRate: { input: any; output: any; }
  Timestamp: { input: any; output: any; }
};
// ...
export type ReleaseDate = {
  __typename?: 'ReleaseDate';
  /**
   * The date the item was first released, for example `1981-12-15`. Depending on
   * the precision, it might be shown as `1981-12`, or `1981-12-15`.
   */
  date: Scalars['String']['output'];
  /** The precision with which the `date` value is known. */
  precision: ReleaseDatePrecision;
};

export enum ReleaseDatePrecision {
  DAY = 'DAY',
  MONTH = 'MONTH',
  YEAR = 'YEAR'
}